Rebuilt repository as UFR
Converted the repository to a MakeFile, and changed the build process such that manual editing in Glyphs is no longer required. It can also build all of the fonts via github actions :).

Both Signika normal and Signika Negative are now building from the standard Signika.glyphs file, and can be done from the command line. I needed to add two instances at either end of the designspace in the glyphs file to make this work, but they can be ignored.

Version string bumped. Hot damn!